% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csmp.R
\name{csmp}
\alias{csmp}
\title{Assess climate-sensitive model performance}
\usage{
csmp(
  airtemp,
  precip,
  objective,
  fixed_var = NULL,
  fixed_return = NULL,
  ref_airtemp = NULL,
  ref_precip = NULL,
  delta_airtemp = NULL,
  delta_precip = NULL,
  return_model = F,
  return_plot_df = F,
  return_plot = F,
  return_pred = F,
  airtemp_limits = c(-3, 1),
  precip_limits = c(250, 580),
  k_val = 3
)
}
\arguments{
\item{airtemp}{Vector of observed annual average air temperatures (째C)}

\item{precip}{Vector of observed annual average precipitation values (째C)}

\item{objective}{Vector of model objective function results (e.g., RMSE, NSE)}

\item{fixed_var}{If desired, a variable to be used for a fixed effect. Should be a vector of observations of this effect that correspond to the positions of other vectors given to the function.}

\item{fixed_return}{If a fixed effect variable is provided, this should be a value of the fixed effect variable that should be used to return answers. Must be a value from within the fixed effect variable vector. If not provided, will take first value in fixed_var}

\item{ref_airtemp}{The reference value of air temperature (째C) in a historical average year. If not provided, will be set to mean of observed vector.}

\item{ref_precip}{The reference value of precipitation (mm) in a historical average year. If not provided, will be set to mean of observed vector.}

\item{delta_airtemp}{The minimum and maximum air temperature values (째C) for comparison. If not supplied, will be set to the 5th and 95th percentiles of the range of observations.}

\item{delta_precip}{The minimum and maximum precipitation values (mm) for comparison. If not supplied, will be set to the 5th and 95th percentiles of the range of observations.}

\item{return_model}{Should the function return the model object? This is useful for examining residuals, but could lead to large outputs.}

\item{return_plot_df}{Should the function return a data frame that can be used to make a contour plot?}

\item{return_plot}{Should the function return a contour plot showing the results? Note, the plot returned is a ggplot() object and can be modified with additional calls to ggplot2 (see examples)}

\item{return_pred}{Should the function return predicted error for the supplied values of airtemp and precip?}

\item{airtemp_limits}{Limits of range on which to predict air temp if returning plot_df}

\item{precip_limits}{Limits of range on which to predict precip if returning plot_df}

\item{k_val}{Number of knots to use in gam - equivalent to k in gam()}
}
\value{
Data frame with climate-sensitive model performance for due to precipitation and air temperature
}
\description{
Assess climate-sensitive model performance
}
\examples{
airtemp <- rnorm(20, mean = 0, sd = 1)
precip <- rnorm(20, mean = 500, sd = 100)
objective <- runif(20, min = 0, max = 1)

ans <- csmp(airtemp, precip, objective, return_plot = TRUE)
ans$stats
ans$contour_plot


}
\author{
Adrienne M. Marshall
}
